+++ lib/ansible/modules/system/user.py
@@ -420,10 +420,7 @@ class User(object):
             cmd.append(self.password)
 
         if self.createhome:
-            cmd.append('-m')
-            if self.local:
-                cmd.append('-M')
-            else:
+            if not self.local:
                 cmd.append('-m')
 
             if self.skeleton is not None: